resolved conversations about POST and PUT for PR
