+++ backend/src/main/java/ru/hh/school/checkupextension/admin/AdminResource.java
@@ -1,17 +1,18 @@
 package ru.hh.school.checkupextension.admin;
 
 import jakarta.inject.Inject;
+import jakarta.validation.Valid;
 import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.CookieParam;
 import jakarta.ws.rs.POST;
+import jakarta.ws.rs.PUT;
 import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
-import ru.hh.school.checkupextension.core.data.dto.admin.EditableProblemDto;
+import ru.hh.school.checkupextension.core.data.dto.contest.ContestProblemDto;
+import ru.hh.school.checkupextension.core.data.entity.Problem;
 
-import static ru.hh.school.checkupextension.utils.constant.CookiesName.USER_TOKEN;
-
-@Path("/admin")
+@Path("/")
 public class AdminResource {
   private final AdminService adminService;
 
@@ -20,13 +21,25 @@ public class AdminResource {
     this.adminService = adminService;
   }
 
+  /**
+   * Создает новую задачу на основе переданной информации в теле запроса.
+   */
   @POST
-  @Path("/problem")
+  @Path("problem")
   @Consumes(MediaType.APPLICATION_JSON)
   @Produces(MediaType.APPLICATION_JSON)
-  public EditableProblemDto createNewProblem(
-      @CookieParam(USER_TOKEN) String userToken,
-      EditableProblemDto problemDto) {
-    return adminService.createNewProblem(userToken, problemDto);
+  public ContestProblemDto createProblem(@Valid Problem problem) {
+    return adminService.createProblem(problem);
+  }
+
+  /**
+   * Метод обновляет задачу с указанным id на основе переданной информации в теле запроса.
+   */
+  @PUT
+  @Path("problem/{id}")
+  @Consumes(MediaType.APPLICATION_JSON)
+  @Produces(MediaType.APPLICATION_JSON)
+  public ContestProblemDto updateProblem(@PathParam("id") long id, Problem problem) {
+    return adminService.updateProblem(id, problem);
   }
 }
