+++ backend/src/main/java/ru/hh/school/checkupextension/contest/ContestResource.java
@@ -1,13 +1,10 @@
 package ru.hh.school.checkupextension.contest;
 
 import jakarta.inject.Inject;
-import jakarta.validation.Valid;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.CookieParam;
-import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
-import jakarta.ws.rs.PUT;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
@@ -16,7 +13,6 @@ import ru.hh.school.checkupextension.core.data.dto.contest.ContestSubmissionDto;
 import ru.hh.school.checkupextension.core.data.dto.contest.ContestProblemDto;
 import ru.hh.school.checkupextension.core.data.dto.contest.ContestSubmissionResultDto;
 import ru.hh.school.checkupextension.core.data.dto.contest.ContestSubmissionShortInfoDto;
-import ru.hh.school.checkupextension.core.data.entity.Problem;
 
 import java.util.List;
 
@@ -44,28 +40,6 @@ public class ContestResource {
   @Produces(MediaType.APPLICATION_JSON)
   public ContestProblemDto getProblem(@PathParam("id")long id) {
     return contestService.getProblem(id);
-  }
-
-  /**
-   * Создает новую задачу на основе переданной информации в теле запроса.
-   */
-  @POST
-  @Path("problem")
-  @Consumes(MediaType.APPLICATION_JSON)
-  @Produces(MediaType.APPLICATION_JSON)
-  public ContestProblemDto createProblem(@Valid Problem problem) {
-    return contestService.createProblem(problem);
-  }
-
-  /**
-   * Метод обновляет задачу с указанным id на основе переданной информации в теле запроса.
-   */
-  @PUT
-  @Path("problem/{id}")
-  @Consumes(MediaType.APPLICATION_JSON)
-  @Produces(MediaType.APPLICATION_JSON)
-  public ContestProblemDto updateProblem(@PathParam("id") long id, Problem problem) {
-    return contestService.updateProblem(id, problem);
   }
 
   @POST
